@@ -1,17 +1,17 @@
 name: simple_dart_web_widgets
 description: Simple dart web widgets.
 version: 0.0.1
-#homepage: https://www.example.com
-#author: mi <email@example.com>
+publish_to: none
+author: Viktor Zakharov  <viktorzahar@gmail.com>
 
 environment:
-  sdk: '^2.7.0'
+  sdk: '^2.12.1'
 
 #dependencies:
 
 dev_dependencies:
-  build_runner: ^1.7.2
-  build_web_compilers: ^2.7.1
+  build_runner: ^1.12.1
+  build_web_compilers: ^2.16.5
 
 web:
   compiler:
