@@ -5,8 +5,6 @@ import '../widgets.dart';
 
 abstract class DialogWindow<T> {
   DialogWindow() {
-    backgroundElement = DivElement();
-
     backgroundElement.style
       ..width = '100%'
       ..height = '100%'
@@ -21,7 +19,6 @@ abstract class DialogWindow<T> {
         closeDialog();
       }
     });
-    dialogElement = DivElement();
     dialogElement.style
       ..position = 'absolute'
       ..display = 'flex'
@@ -29,47 +26,45 @@ abstract class DialogWindow<T> {
       ..padding = '5px'
       ..opacity = '1'
       ..flexDirection = 'column';
-    buttonsPanelElement = DivElement();
     buttonsPanelElement.style
       ..width = '100%'
       ..flexDirection = 'row';
 
-    captionElement = DivElement();
     dialogElement.children.add(captionElement);
     dialogElement.children.add(buttonsPanelElement);
     dialogElement.children.add(createDialogContent().nodeRoot);
   }
 
-  DivElement backgroundElement;
-  DivElement dialogElement;
-  DivElement captionElement;
-  DivElement buttonsPanelElement;
-  Completer<T> completer;
-  Function() onCloseListener;
+  DivElement backgroundElement = DivElement();
+  DivElement dialogElement = DivElement();
+  DivElement captionElement = DivElement();
+  DivElement buttonsPanelElement = DivElement();
+  Completer<T>? completer;
+  Function()? onCloseListener;
   bool closable = true;
 
   Future showDialog() {
     completer = Completer<T>();
-    final bodyElement = window.document.querySelector('body');
+    final bodyElement = window.document.querySelector('body')!;
     captionElement.text = caption();
     bodyElement.children.add(backgroundElement);
     bodyElement.children.add(dialogElement);
 
     center();
-    return completer.future;
+    return completer!.future;
   }
 
   void closeDialog() {
     backgroundElement.remove();
     dialogElement.remove();
     if (onCloseListener != null) {
-      onCloseListener();
+      onCloseListener!();
     }
   }
 
   void center() {
-    final windowWidth = window.innerWidth;
-    final windowHeight = window.innerHeight;
+    final windowWidth = window.innerWidth!;
+    final windowHeight = window.innerHeight!;
     final dialogWidth = dialogElement.clientWidth;
     final dialogHeight = dialogElement.clientHeight;
     final x = (windowWidth - dialogWidth) / 2;
